collabs: filter duplicates
